requirements.txt
pyqt5
torch
torchvision
loguru
numpy
Pillow
opencv-python
matplotlib